refactor(sbi-testing): split thread context switch per architecture
Restructure the thread module into a shared part (context type, safe
API, layout contract) and per-architecture naked switchers: 8-byte
slots with sd/ld on RV64, 4-byte slots with sw/lw on RV32.

- pin the Thread layout against the assembly with compile-time offset
  assertions; a field reorder now fails the build instead of silently
  restoring wrong registers
- take the stack pointer in Thread::new(entry, sp) instead of a
  separate mutation, and debug-assert register indices
- keep the call site 16-byte ABI-aligned when saving ra
- rewrite comments in English and document the layout contract

The switch bodies match the protocol previously exercised on qemu for
both architectures; the ra save frame is new and remains to be
re-exercised in the next qemu boot run.
